US Ta-The M: Deletions
That Fish Place is just a pet shop. The Claremont Village is a shopping
mall. The Golden Lamb is a hotel and restaurant.

Th' Can, The Anime Planet I can't find any sign of.

The Victorian Festival does appear to have happened, but I can't find
anything about it in recent years.

The Insectarium and the Bee Hive appear to have closed. (The former with
extreme prejudice -- they left $1m in mortgage arrears, and trashed the
place on the way out).

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>