ZZ: Deletions
Abu Qir is a whole town.

I can't find a reference to the bunker entrance in Yerevan. There's a
game cafe with a bunch of PS5s called the Bunker, but the location doesn't
match up.

The Atlas Mountains Trek is similar to the Australian travel company --
organised multi-day tours, not really a day trip.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>